+++ packages/lucid/test/emulator/service.ts
@@ -191,3 +191,24 @@ export const evaluateAContractWithDatum = Effect.gen(function* ($) {
     .completeProgram();
   return signBuilder;
 }).pipe(Effect.flatMap(handleSignSubmitWithoutValidation), withLogRetry);
+
+export const compose = Effect.gen(function* ($) {
+  const { user } = yield* EmulatorUser;
+  const scriptAddress = validatorToAddress("Custom", alwaysSucceedScript);
+  const txCompA = user
+    .newTx()
+    .pay.ToContract(
+      scriptAddress,
+      { kind: "asHash", value: Data.to("31313131") },
+      { lovelace: 5000000n },
+    );
+  const txCompB = user
+    .newTx()
+    .pay.ToContract(
+      scriptAddress,
+      { kind: "inline", value: Data.to("31313131") },
+      { lovelace: 5000000n },
+    );
+  const signBuilder = yield* txCompA.compose(txCompB).completeProgram();
+  return signBuilder;
+}).pipe(Effect.flatMap(handleSignSubmitWithoutValidation), withLogRetry);