+++ packages/lucid/test/onchain-preview.test.ts
@@ -14,6 +14,7 @@ import * as MintBurnEndpoints from "./specs/mint-burn.js";
 import * as ParametrizedEndpoints from "./specs/hello-params.js";
 import * as TxChain from "./specs/tx-chaining.js";
 import * as MetadataEndpoint from "./specs/metadata.js";
+import * as WalletEndpoint from "./specs/wallet.js";
 
 describe.sequential("Onchain testing", () => {
   test("TxChain", async () => {
@@ -318,5 +319,16 @@ describe.sequential("Onchain testing", () => {
     );
     const exit = await Effect.runPromiseExit(program);
     expect(exit._tag).toBe("Failure");
+  });
+
+  //Helps recycle utxos when the coin selection algorithm is set to Largest first
+  test("recycleUTxOs", async () => {
+    const program = pipe(
+      WalletEndpoint.recycleUTxOs,
+      Effect.provide(User.layer),
+      Effect.provide(NetworkConfig.layerPreview),
+    );
+    const exit = await Effect.runPromiseExit(program);
+    expect(exit._tag).toBe("Success");
   });
 });