+++ packages/lucid/test/specs/services.ts
@@ -71,7 +71,13 @@ const makeUser = Effect.gen(function* ($) {
   user.selectWallet.fromSeed(networkConfig.WALLET_SEED);
   yield* pipe(
     Effect.promise(() => user.wallet().address()),
-    Effect.flatMap((address) => Effect.log("init wallet", address)),
+    Effect.flatMap((address) => Effect.log(`Wallet : ${address}`)),
+  );
+  yield* pipe(
+    Effect.promise(() => user.wallet().getUtxos()),
+    Effect.flatMap((utxos) =>
+      Effect.log(`Total Wallet UTxOs: ${utxos.length}`),
+    ),
   );
   return {
     user,