+++ packages/lucid/vitest.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
     pool: "forks",
     reporters: "verbose",
     include: ["./test/**/*.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"],
-    testTimeout: 300_000,
-    bail: 2,
+    testTimeout: 420_000,
+    bail: 3,
   },
 });