+++ packages/provider/src/kupmios/kupmios.ts
@@ -16,7 +16,7 @@ import {
   Unit,
   UTxO,
 } from "@lucid-evolution/core-types";
-import { fromUnit } from "@lucid-evolution/utils";
+import { applyDoubleCborEncoding, fromUnit } from "@lucid-evolution/utils";
 import * as S from "@effect/schema/Schema";
 import { Effect, pipe, Array as _Array, Schedule, Data } from "effect";
 import * as KupmiosSchema from "./schema.js";
@@ -199,7 +199,7 @@ export class Kupmios implements Provider {
   }
 
   async awaitTx(txHash: TxHash, checkInterval = 20000): Promise<boolean> {
-    const pattern = `${this.kupoUrl}/matches/*@${txHash}?unspent`;
+    const pattern = `${this.kupoUrl}/matches/*@${txHash}?spent`;
     const schema = S.Array(KupmiosSchema.KupoUTxO).annotations({
       identifier: "Array<KupmiosSchema.KupoUTxO>",
     });
@@ -329,7 +329,7 @@ const getDatumEffect = (
   datum_type: KupmiosSchema.KupoUTxO["datum_type"],
   datum_hash: KupmiosSchema.KupoUTxO["datum_hash"],
 ): Effect.Effect<
-  string | null,
+  string | undefined,
   HttpClientError | ParseError | TimeoutException | NoSuchElementException
 > =>
   Effect.gen(function* () {
@@ -341,7 +341,7 @@ const getDatumEffect = (
         Effect.map((result) => result.datum),
         Effect.timeout(10_000),
       );
-    } else return null;
+    } else return undefined;
   });
 
 const getScriptEffect = (
@@ -359,15 +359,21 @@ const getScriptEffect = (
         Effect.map(({ language, script }) => {
           switch (language) {
             case "native":
-              return { type: "Native", script } satisfies Script;
+              return undefined;
             case "plutus:v1":
-              return { type: "PlutusV1", script } satisfies Script;
+              return {
+                type: "PlutusV1",
+                script: applyDoubleCborEncoding(script),
+              } satisfies Script;
             case "plutus:v2":
-              return { type: "PlutusV2", script } satisfies Script;
+              return {
+                type: "PlutusV2",
+                script: applyDoubleCborEncoding(script),
+              } satisfies Script;
           }
         }),
       );
-    } else return null;
+    } else return undefined;
   });
 
 const toAssets = (value: KupmiosSchema.KupoUTxO["value"]): Assets => {
@@ -473,7 +479,7 @@ const kupmiosUtxosToUtxos = (
             outputIndex: utxo.output_index,
             address: utxo.address,
             assets: toAssets(utxo.value),
-            datumHash: utxo.datum_type === "hash" ? utxo.datum_hash : null,
+            datumHash: utxo.datum_type === "hash" ? utxo.datum_hash : undefined,
             datum: datum,
             scriptRef: script,
           }),