+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lucid-evolution/wallet
 
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [[`344b8cb`](https://github.com/Anastasia-Labs/lucid-evolution/commit/344b8cbf376d59fd1b6fbf9a17ab892559bf325f)]:
+  - @lucid-evolution/utils@0.1.35
+
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [[`803086b`](https://github.com/Anastasia-Labs/lucid-evolution/commit/803086be99793f334b8c82eb2edd739b3e57ba37)]:
+  - @lucid-evolution/utils@0.1.34
+
 ## 0.1.37
 
 ### Patch Changes